Refactor EventCard out, figure out link to event
